Move in-browser code out of main function, and into standalone BrowserInterface functions
