+++ lib/generate-critical-css.js
@@ -2,7 +2,7 @@ const CSSFileSet = require( './css-file-set' );
 const { removeIgnoredPseudoElements } = require( './ignored-pseudo-elements' );
 const { minifyCss } = require( './minify-css' );
 const { SuccessTargetError, EmptyCSSError } = require( './errors' );
-const { BrowserInterface } = require( './browser-interface' );
+const BrowserInterface = require( './browser-interface' );
 
 /**
  * Collate and return a CSSFileSet object describing all the CSS files used by
@@ -101,19 +101,7 @@ async function generateCriticalCSS( {
 			const pageSelectors = await browserInterface.runInPage(
 				url,
 				null,
-				( innerWindow, selText, trimmed ) => {
-					return selText.filter( ( selector ) => {
-						try {
-							return !! innerWindow.document.querySelector(
-								trimmed[ selector ]
-							);
-						} catch ( err ) {
-							// Ignore invalid selectors.
-							return false;
-						}
-					} );
-				},
-				selectorText,
+				BrowserInterface.innerFindMatchingSelectors,
 				trimmedSelectors
 			);
 
@@ -129,38 +117,9 @@ async function generateCriticalCSS( {
 				const pageAboveFold = await browserInterface.runInPage(
 					url,
 					size,
-					( innerWindow, pageSel, trimmed ) => {
-						const isAboveFold = ( element ) => {
-							const originalClearStyle =
-								element.style.clear || '';
-							element.style.clear = 'none';
-
-							const rect = element.getBoundingClientRect();
-
-							element.style.clear = originalClearStyle;
-
-							return rect.top < innerWindow.innerHeight;
-						};
-
-						return pageSel.filter( ( s ) => {
-							if ( '*' === trimmed[ s ] ) {
-								return true;
-							}
-
-							const matches = innerWindow.document.querySelectorAll(
-								trimmed[ s ]
-							);
-							for ( const match of matches ) {
-								if ( isAboveFold( match ) ) {
-									return true;
-								}
-							}
-
-							return false;
-						} );
-					},
-					pageSelectors,
-					trimmedSelectors
+					BrowserInterface.innerFindAboveFoldSelectors,
+					trimmedSelectors,
+					pageSelectors
 				);
 
 				pageAboveFold.forEach( ( s ) => aboveFoldSelectors.add( s ) );
